backend/web/requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
MySQL-python==1.2.5
virtualenv==15.1.0
